pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "burnin-subtitle-checker"
version = "0.2.0"
description = "CLI tool for flagging mismatches between audio dialogue and burned-in subtitles."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "PlanetRead community" }
]
keywords = ["subtitles", "ocr", "asr", "whisper", "tesseract", "hindi", "kannada"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Video",
  "Topic :: Text Processing :: Linguistic",
]
dependencies = []

[project.optional-dependencies]
similarity = ["rapidfuzz>=3.6", "jiwer>=3.0"]
ocr-preprocess = ["opencv-python-headless>=4.8", "numpy>=1.24"]
ocr-easy = ["easyocr>=1.7"]
asr = ["openai-whisper>=20231117"]
asr-fast = ["faster-whisper>=1.0"]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]
all = [
  "rapidfuzz>=3.6",
  "jiwer>=3.0",
  "opencv-python-headless>=4.8",
  "numpy>=1.24",
  "openai-whisper>=20231117",
  "faster-whisper>=1.0",
  "easyocr>=1.7",
]

[project.scripts]
burnsub = "burnin_subtitle_checker.cli:main"
burnin-subtitle-checker = "burnin_subtitle_checker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
burnin_subtitle_checker = ["templates/*.html", "assets/*.css"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
